data/requirements.txt
# --- Web API & Database ---
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9
psycopg2-binary==2.9.9
boto3==1.34.0
prometheus-client==0.20.0
prometheus-fastapi-instrumentator==6.1.0

# --- Online Feature Computation & Image Processing ---
numpy==1.26.4
Pillow==10.2.0
pycocotools==2.0.7
scikit-learn==1.7.2
alibi-detect==0.12.0

# --- Synthetic Data Generation ---
requests==2.31.0
faker==24.0.0
tqdm==4.66.2

# --- Batch Pipeline & Lakehouse ---
pyspark==3.5.0
pyiceberg[s3fs,sql-postgres]==0.11.1
pyarrow==15.0.0
mlflow==3.1.4
